ai-model/requirements.txt
absl-py==2.1.0
apt-clone==0.2.1
apturl==0.5.2
arandr==0.1.10
astunparse==1.6.3
async-timeout==4.0.1
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
catfish==4.16.4
ccsm==0.9.14.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
eyeD3==0.8.10
filelock==3.6.0
flatbuffers==25.2.10
gast==0.6.0
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
imageio==2.36.1
imageio-ffmpeg==0.5.1
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
keras==3.8.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
louis==3.20.0
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
menulibre==2.2.2
ml-dtypes==0.4.1
more-itertools==8.10.0
moviepy==1.0.3
mugshot==0.4.3
mutagen==1.45.1
namex==0.0.8
netaddr==0.8.0
netifaces==0.11.0
numpy==2.0.2
oauthlib==3.2.0
onboard==1.4.1
opt_einsum==3.4.0
optree==0.14.0
packaging==21.3
pexpect==4.8.0
pillow==10.4.0
proglog==0.1.10
protobuf==5.29.3
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-magic==0.4.24
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rich==13.9.4
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
speedtest-cli==2.1.3
ssh-import-id==5.11
systemd-python==234
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tldextract==3.1.2
tqdm==4.67.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
websockets==9.1
Werkzeug==3.1.3
wrapt==1.17.2
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
yt-dlp==2022.4.8
zipp==1.0.0
